# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "اغتنم API Documentation",
        "description": "API documentation for اغتنم application",
        "contact": {
            "name": "Admin",
            "email": "admin@e8tanem.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/videos": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "List all active videos",
                "description": "Returns a paginated list of active videos with pagination metadata",
                "operationId": "2090c29a8b2a49778e2de02f54ec7606",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination (starts from 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (default: 20)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term to filter videos by title or description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Field to sort by (created_at, views_count, likes_count, comments_count)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at",
                            "enum": [
                                "created_at",
                                "views_count",
                                "likes_count",
                                "comments_count"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of active videos with pagination metadata",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Sample Video"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Video description"
                                                    },
                                                    "video_url": {
                                                        "type": "string",
                                                        "example": "/storage/videos/sample-video.mp4"
                                                    },
                                                    "thumbnail_url": {
                                                        "type": "string",
                                                        "example": "/storage/thumbnails/sample-thumbnail.jpg"
                                                    },
                                                    "views_count": {
                                                        "type": "integer",
                                                        "example": 100
                                                    },
                                                    "likes_count": {
                                                        "type": "integer",
                                                        "example": 50
                                                    },
                                                    "comments_count": {
                                                        "type": "integer",
                                                        "example": 25
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "John Doe"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://e8tanem.com/api/videos?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://e8tanem.com/api/videos?page=5"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "label": {
                                                        "type": "string",
                                                        "example": "&laquo; Previous"
                                                    },
                                                    "active": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://e8tanem.com/api/videos?page=2",
                                            "nullable": true
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://e8tanem.com/api/videos"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": "http://e8tanem.com/api/videos?page=1",
                                            "nullable": true
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 100
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/videos/{id}": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get video details",
                "operationId": "d5190024e38b00df6680e59ca8a482b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the video",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Sample Video"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Video description"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "https://example.com/video.mp4"
                                        },
                                        "views_count": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "likes_count": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "comments_count": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "is_active": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Video not found"
                    }
                }
            }
        },
        "/api/videos/{id}/stream": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Stream video content",
                "description": "Stream the video content with support for range requests and caching for mobile optimization",
                "operationId": "66359a8a5befba239ec45b6a7631d6b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the video to stream",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Range",
                        "in": "header",
                        "description": "Range header for partial content requests (e.g., bytes=0-1000)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Full video content stream",
                        "headers": {
                            "Content-Type": {
                                "description": "video/mp4",
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "Accept-Ranges": {
                                "description": "bytes",
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "Content-Length": {
                                "description": "Size of the video in bytes",
                                "schema": {
                                    "type": "integer"
                                }
                            },
                            "Cache-Control": {
                                "description": "public, max-age=86400",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "206": {
                        "description": "Partial video content (for range requests)",
                        "headers": {
                            "Content-Type": {
                                "description": "video/mp4",
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "Content-Range": {
                                "description": "bytes start-end/size",
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "Content-Length": {
                                "description": "Size of the requested range in bytes",
                                "schema": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Video not found or file not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UnauthorizedError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "NotFoundError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}